requirements.txt
flask 
selenium
webdriver_manager
bs4
requests    
pillow 
pytesseract
opencv-python